Sync code

+++ repeater.py
@@ -16,14 +16,15 @@ s3 = boto3.client(
     aws_access_key_id=env["ACCESS_KEY"],
     aws_secret_access_key=env["SECRET_KEY"],
 )
+
 directory_path = {
-    "hi": "/home/hankookilbo/",  # 한국일보
-    "mk": "/home/mk/",  # 매일경제
-    "fn": "home/fnnews",  # 파이낸셜
-    "hn": "/home/hani/",  # 한겨례
-    "ja": "/home/joongang/",  # 중앙일보
-    "kh": "/home/khan/",  # 경향신문
-    "hk": "/home/hankyung/",  # 한국경제
+    "hi": "/home/hankookilbo",  # 한국일보
+    "mk": "/home/mk",  # 매일경제
+    "fn": "/home/fnnews",  # 파이낸셜
+    "hn": "/home/hani",  # 한겨례
+    "ja": "/home/joongang",  # 중앙일보
+    "kh": "/home/khan",  # 경향신문
+    "hk": "/home/hankyung",  # 한국경제
 }
 
 # def download_file(ftp, filename, filepath):
@@ -100,26 +101,29 @@ def process_files(press_name):
     for filename in file_list:
         source_path = os.path.join(file_direcotry, filename)
         destination_path = os.path.join(process_directory, filename)
+
+        origin_file_exist = os.path.exists(destination_path)
+
         modifited_time = os.stat(source_path).st_mtime
         now = datetime.now(kst).timestamp()
-        if (now - modifited_time) < 60 * 10:
+        if not origin_file_exist and (now - modifited_time) < 60 * 10:
             copy2(source_path, destination_path)
 
             print("COPY", filename)
 
             object_name = f"origin_news/{press_name}/{filename}"
             upload_file_to_s3(os.path.join(destination_path), object_name)
-            os.remove(destination_path)
 
             print("UPLOAD", filename, object_name)
 
 
 def main():
-    press_name = env["PRESS_NAME"]
+    process_press_list = ["hi"]
     need_sync_press = ["mk", "fn"]
     # if press_name in need_sync_press:
     #     sync_press()
-    process_files(press_name)
+    for press_name in process_press_list:
+        process_files(press_name)
 
 
 # Run the main function
